requirements.txt
fastapi==0.104.1
uvicorn[standard]
pymongo==4.5.0
pytz==2023.3.post1
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
# uvicorn==0.24.0.post